% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testUtitities.R
\name{benchmarkCore}
\alias{benchmarkCore}
\title{Private function for benchmark}
\usage{
benchmarkCore(populationCount, dimentions, fnc, runs, initSpawnArea,
  maxEvaluations, c, d)
}
\description{
Private function for benchmark
}

