% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testUtitities.R
\name{easomFunct}
\alias{easomFunct}
\title{Flat function with one deep, but small peek, implemented only for 2 dimentions}
\usage{
easomFunct(z)
}
\arguments{
\item{x}{A point to be evaluated}
}
\value{
quality of the point
}
\description{
Flat function with one deep, but small peek, implemented only for 2 dimentions
}

