% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateStartPoints}
\alias{generateStartPoints}
\title{Start point generator function}
\usage{
generateStartPoints(popCount, dim, min, max)
}
\arguments{
\item{popCount}{Swarm size}

\item{dim}{Dimention count}

\item{min}{lower bound on coordinates generation}

\item{max}{upper bound on coordinates generation}
}
\value{
history The new history.
}
\description{
Start point generator function
}

