% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testUtitities.R
\name{griewankFunct}
\alias{griewankFunct}
\title{Very spiky function, best results with larger search area (-600 600)}
\usage{
griewankFunct(z)
}
\arguments{
\item{x}{A point to be evaluated}
}
\value{
quality of the point
}
\description{
Very spiky function, best results with larger search area (-600 600)
}

