% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initialization}
\alias{initialization}
\title{Default history initialization}
\usage{
initialization(startPoints)
}
\arguments{
\item{startPoints}{The start points.}
}
\value{
history The new history.
}
\description{
Default history initialization
}

