% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testUtitities.R
\name{jongFunct}
\alias{jongFunct}
\title{A simple quadratic function}
\usage{
jongFunct(x)
}
\arguments{
\item{x}{A point to be evaluated}
}
\value{
quality of the point
}
\description{
A simple quadratic function
}

