% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaheuristic_pattern.r
\name{metaheuristicRun}
\alias{metaheuristicRun}
\title{Main function that runs metaheuristic PSO algorithm implementation.}
\usage{
metaheuristicRun(initialization, startPoints, termination, evaluation,
  initSpawnArea = 10, maxEvaluations = 100,
  localBestVsGlobalBestRatio = 0.75, explorationVsExploitaitionRatio = 0.9)
}
\arguments{
\item{initialization}{A method that constructs history from start points}

\item{startPoints}{A group of start Points}

\item{termination}{A method that stops algorithm after certain condition is met}

\item{evaluation}{A method that counts the quality of the points}

\item{initSpawnArea}{A radius according to which initial velocities of particles are generated}

\item{maxEvaluations}{Maximum number of evaluations}

\item{localBestVsGlobalBestRatio}{A number between 0 and 1 telling how strongly particle will be pulled against best known points}

\item{explorationVsExploitaitionRatio}{A number between 0 and 1 telling if particles will explore or exploit more}
}
\value{
history of the run
}
\description{
Main function that runs metaheuristic PSO algorithm implementation.
}

