% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testUtitities.R
\name{minHistory}
\alias{minHistory}
\title{Finds the best point in history}
\usage{
minHistory(x)
}
\arguments{
\item{x}{A history}
}
\value{
best point in history
}
\description{
Finds the best point in history
}

