% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testUtitities.R
\name{plotQuality}
\alias{plotQuality}
\title{Plots the change of quality in time}
\usage{
plotQuality(history, fName)
}
\arguments{
\item{history}{A history for chaining.}
}
\description{
Plots the change of quality in time
}

