% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testUtitities.R
\name{rosenbrockFunct}
\alias{rosenbrockFunct}
\title{Function with easy to find local optimum, but hard global optimum to converge too, only 2 dimentions}
\usage{
rosenbrockFunct(z)
}
\arguments{
\item{x}{A point to be evaluated}
}
\value{
quality of the point
}
\description{
Function with easy to find local optimum, but hard global optimum to converge too, only 2 dimentions
}

