% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{terminateByEvaluationsCount}
\alias{terminateByEvaluationsCount}
\title{Terminates execution of algorithm after number of evaluatons exceeds certain treshold}
\usage{
terminateByEvaluationsCount(history, model)
}
\arguments{
\item{history}{A history.}

\item{model}{A model.}
}
\value{
decision if algorithm has to be stopped.
}
\description{
Terminates execution of algorithm after number of evaluatons exceeds certain treshold
}

