% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{termination}
\alias{termination}
\title{Adapter for terminateByEvaluationsCount so it can be used in algorithm
Uses parameter app.maxEvaluations as limit of evaluations}
\usage{
termination(history, model)
}
\arguments{
\item{history}{A history.}

\item{model}{A model.}
}
\value{
decision if algorithm has to be stopped.
}
\description{
Adapter for terminateByEvaluationsCount so it can be used in algorithm
Uses parameter app.maxEvaluations as limit of evaluations
}

