% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testUtitities.R
\name{trace}
\alias{trace}
\title{Plots the graph with the function trace from history}
\usage{
trace(z, drawLines = FALSE)
}
\arguments{
\item{z}{A history}

\item{drawLines}{- should draw lines or not}
}
\value{
quality of best point
}
\description{
Plots the graph with the function trace from history
}

