% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC.R
\name{abc}
\alias{abc}
\title{Implementation of Artificial Bee Colony(ABC) algorithm
for the purposes of ALHE class on Warsaw University of Technology}
\usage{
abc(goal, dim, pars)
}
\arguments{
\item{goal, }{f(x) - funkcja celu obliczana dla argumentu x}

\item{dim, }{wymiarowosc dla ktorej prowadzone beda obliczenia}

\item{pars, }{obiekt z pozostalymi parametrami metody:
- NP: ilosc pszczol,
- FoodNumber: ilosc zrodl jedzenia
- lb:  dolna granica zasiegu przeszukiwania - dla przyspieszenia
- ub:  gorna granica parametrow - dla przyspieszenia
- limit: zrodlo jedzenia, ktore nie moze byc ulepszone w tylu probach jest porzucane przez pszczoly
- maxCycle: maksymalna ilosc cykli(iteracji) - "stop" kryterium zatrzymania
- optbin:  TRUE jezeli chcemy optymalizowac binarnie [1,0]
- critical: ograniczenie na niezmiennosc rozwiazania - dla przyspieszenia}
}
\value{
res - wynik dzia<U+D762>ania metody, przy czym:
    - res$x to wektor odpowiadaj<U+D6F3>cy najlepszemu rozwi<U+D6F3>zaniu,
    - res$y to warto<U+D784><U+D6F6> f. celu dla tego wektora
}
\description{
Package: alhe
}
\examples{
abc(function(x) sum(x^2), 5, list(NP=40,lb=-100, ub=100, critical=100))

}
\author{
Volodymyr Ostruk \email{ostruk@outlook.com}
Finds a local minimums of the function(aka "the best food sources")
}
\references{
\url{https://en.wikipedia.org/wiki/Artificial_bee_colony_algorithm}

Dervis KARABOGA; AN IDEA BASED ON HONEY BEE SWARM FOR NUMERICAL OPTIMIZATION; TECHNICAL REPORT-TR06, OCTOBER, 2005
}

