% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC.R
\name{abc}
\alias{abc}
\title{Implementation of Artificial Bee Colony(ABC) algorithm
for the purposes of ALHE class on Warsaw University of Technology}
\usage{
abc(goal, dim, par, ..., NP = 40, FoodNumber = NP/2, lb = -Inf,
  ub = +Inf, limit = 100, maxCycle = 1000, optbin = FALSE,
  criter = 50)
}
\arguments{
\item{goal}{f(x) f(x) - funkcja celu obliczana dla argumentu x}

\item{dim}{wymiarowosc dla ktorej prowadzone beda obliczenia}

\item{par}{# wektor z parametrami dla danej konkretnej funkcji}
}
\value{
res - wynik działania metody, przy czym:
 - res$x to wektor odpowiadający najlepszemu rozwiązaniu,
 -res$y to wartość f. celu dla tego wektora
}
\description{
Finds a local minimums of the function(aka "the best food sources")
}
\examples{
abc(function(x) {100 * (x[2] - x[1]^2)^2 + (1 - x[1])^2}, 2, c(1,1), lb=-2.028, ub=2.028, criter=200)

}

